move: 공통 기능 패키지 정리
