+++ src/main/java/gift/global/config/FilterConfig.java
@@ -1,4 +1,4 @@
-package gift.config;
+package gift.global.config;
 
 import gift.auth.security.JwtFilter;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
