+++ src/main/java/gift/global/error/GlobalExceptionAdvice.java
@@ -1,4 +1,4 @@
-package gift.error;
+package gift.global.error;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
