+++ src/test/java/gift/api/ProductControllerTest.java
@@ -3,8 +3,8 @@ package gift.api;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import gift.auth.security.JwtFilter;
 import gift.auth.security.JwtUtil;
-import gift.error.CustomException;
-import gift.error.ErrorCode;
+import gift.global.error.CustomException;
+import gift.global.error.ErrorCode;
 import gift.product.api.ProductController;
 import gift.product.application.ProductService;
 import gift.product.dto.ProductRequest;
